// src/app/models/sending.ts
import {SuperCategory} from "./supercategory";

export class Sending {
  id: number;
  name: string;
  surnames: string;
  dni: string;
  phone: string;
  address: string;
  address_detail: string;
  notes: string;
  country: string;
  zip: string;
  city: string;
  province: string;
  email: string;
  carrier_id: string;
  service: string;
  carrier: string;
  rate_id: string;
  rate: string;
  delivery_days: string;
  NewOrderID: string;
  LabelURL: string;
  status: string;
  date: string;
  ref: string;
  track_num: string;
  weight: string;
  length: string;
  width: string;
  height: string;
}